refactor: 🔨alertable 파일 분리(#142)
