+++ ios/IssueTracker/IssueTracker/Login/ViewController/Alertable.swift
@@ -1,0 +1,28 @@
+//
+//  Alertable.swift
+//  IssueTracker
+//
+//  Created by Lia on 2021/06/25.
+//
+
+import UIKit
+
+protocol Alertable { }
+
+extension Alertable where Self: UIViewController {
+    
+    func showAlert(title: String = "", message: String, preferredStyle: UIAlertController.Style = .alert, completion: (() -> Void)? = nil) {
+        let alert = UIAlertController(title: title, message: message, preferredStyle: .alert)
+        alert.addAction(UIAlertAction(title: "OK", style: UIAlertAction.Style.default, handler: nil))
+        self.present(alert, animated: true, completion: completion)
+    }
+    
+    func showAlertWithDimiss(title: String = "", message: String, preferredStyle: UIAlertController.Style = .alert, completion: (() -> Void)? = nil) {
+        let alert = UIAlertController(title: title, message: message, preferredStyle: .alert)
+        alert.addAction(UIAlertAction(title: "OK", style: UIAlertAction.Style.default, handler: { _ in
+            self.dismiss(animated: true, completion: nil)
+        }))
+        self.present(alert, animated: true, completion: completion)
+    }
+    
+}
