+++ ios/IssueTracker/IssueTracker/Login/ViewController/LoginViewController.swift
@@ -64,21 +64,3 @@ extension LoginViewController: ASWebAuthenticationPresentationContextProviding, 
     }
     
 }
-
-protocol Alertable { }
-
-extension Alertable where Self: UIViewController {
-    func showAlert(title: String = "", message: String, preferredStyle: UIAlertController.Style = .alert, completion: (() -> Void)? = nil) {
-        let alert = UIAlertController(title: title, message: message, preferredStyle: .alert)
-        alert.addAction(UIAlertAction(title: "OK", style: UIAlertAction.Style.default, handler: nil))
-        self.present(alert, animated: true, completion: completion)
-    }
-    
-    func showAlertWithDimiss(title: String = "", message: String, preferredStyle: UIAlertController.Style = .alert, completion: (() -> Void)? = nil) {
-        let alert = UIAlertController(title: title, message: message, preferredStyle: .alert)
-        alert.addAction(UIAlertAction(title: "OK", style: UIAlertAction.Style.default, handler: { _ in
-            self.dismiss(animated: true, completion: nil)
-        }))
-        self.present(alert, animated: true, completion: completion)
-    }
-}
